+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 ### Fixed
+- [2026-05-06] Rewrote `search_cars` from dead `/nextjs-api/proxy` JSON endpoint to HTML scraping of `cars.ksl.com/search/...` pages. KSL deprecated the JSON API; now uses same scraping approach as classifieds.
 
 ### Removed